# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "{{.Description}}",
        "title": "{{.Title}}",
        "contact": {},
        "license": {},
        "version": "{{.Version}}"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
        "/user": {
            "post": {
                "description": "增加用户",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "增加用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "代理人ID",
                        "name": "agent_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "姓名",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "年龄",
                        "name": "age",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/docs.ResponseOk"
                            }
                        }
                    },
                    "400": {
                        "description": "bad parameters",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.ResponseInvalidParam"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.ResponseServerErr"
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "description": "根据用户ID,获取用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "根据用户ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/docs.ResponseUser"
                            }
                        }
                    },
                    "400": {
                        "description": "bad parameters",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.ResponseInvalidParam"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.ResponseServerErr"
                        }
                    }
                }
            },
            "put": {
                "description": "更新用户信息",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "更新用户信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "姓名",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "年龄",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/docs.ResponseOk"
                            }
                        }
                    },
                    "400": {
                        "description": "bad parameters",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.ResponseInvalidParam"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.ResponseServerErr"
                        }
                    }
                }
            },
            "delete": {
                "description": "根据用户ID,删除用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/docs.ResponseOk"
                            }
                        }
                    },
                    "400": {
                        "description": "bad parameters",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.ResponseInvalidParam"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.ResponseServerErr"
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "根据条件筛选用户",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户列表接口",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "姓名",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "年龄",
                        "name": "age",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.ResponseUsers"
                        }
                    },
                    "400": {
                        "description": "bad parameters",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.ResponseInvalidParam"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.ResponseServerErr"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "docs.ResponseInvalidParam": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "data": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "param is not valid"
                }
            }
        },
        "docs.ResponseOk": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "string",
                    "example": "success"
                },
                "message": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "docs.ResponseServerErr": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "data": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "internal server error"
                }
            }
        },
        "docs.ResponseUser": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/docs.User"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "docs.ResponseUsers": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/docs.User"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "docs.User": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}